+++ Controllers/UsuariosController.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -21,44 +22,48 @@ namespace pomodoro.Controllers
         }
 
         // GET: api/Usuarios
-        [HttpGet]
-        public async Task<ActionResult<IEnumerable<Usuario>>> GetUsuarios()
-        {
-          if (_context.Usuarios == null)
-          {
-              return NotFound();
-          }
-            return await _context.Usuarios.ToListAsync();
-        }
+        // [HttpGet]
+        // public async Task<ActionResult<IEnumerable<Usuario>>> GetUsuarios()
+        // {
+        //   if (_context.Usuarios == null)
+        //   {
+        //       return NotFound();
+        //   }
+        //     return await _context.Usuarios.ToListAsync();
+        // }
 
-        // GET: api/Usuarios/5
-        [HttpGet("{id}")]
-        public async Task<ActionResult<Usuario>> GetUsuario(long id)
-        {
-          if (_context.Usuarios == null)
-          {
-              return NotFound();
-          }
-            var usuario = await _context.Usuarios.FindAsync(id);
+        // // GET: api/Usuarios/5
+        // [HttpGet("{id}")]
+        // public async Task<ActionResult<Usuario>> GetUsuario(long id)
+        // {
+        //   if (_context.Usuarios == null)
+        //   {
+        //       return NotFound();
+        //   }
+        //     var usuario = await _context.Usuarios.FindAsync(id);
 
-            if (usuario == null)
-            {
-                return NotFound();
-            }
+        //     if (usuario == null)
+        //     {
+        //         return NotFound();
+        //     }
 
-            return usuario;
-        }
+        //     return usuario;
+        // }
 
         // PUT: api/Usuarios/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{id}")]
+        [Authorize]
         public async Task<IActionResult> PutUsuario(long id, Usuario usuario)
         {
             if (id != usuario.UsuarioId)
             {
                 return BadRequest();
             }
-
+                    
+            if(!(_context.Usuarios?.Any(x => x.Login == User.Identity.Name)??false)){
+                return Unauthorized();
+            }
             _context.Entry(usuario).State = EntityState.Modified;
 
             try
@@ -83,6 +88,7 @@ namespace pomodoro.Controllers
         // POST: api/Usuarios
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPost]
+        [AllowAnonymous]
         public async Task<ActionResult<Usuario>> PostUsuario(Usuario usuario)
         {
           if (_context.Usuarios == null)
@@ -97,6 +103,7 @@ namespace pomodoro.Controllers
 
         // DELETE: api/Usuarios/5
         [HttpDelete("{id}")]
+        [Authorize]
         public async Task<IActionResult> DeleteUsuario(long id)
         {
             if (_context.Usuarios == null)
@@ -108,7 +115,9 @@ namespace pomodoro.Controllers
             {
                 return NotFound();
             }
-
+            if(!(_context.Usuarios?.Any(x => x.Login == User.Identity.Name && x.UsuarioId == id)??false)){
+                return Unauthorized();
+            }
             _context.Usuarios.Remove(usuario);
             await _context.SaveChangesAsync();
